requirements.txt
streamlit
openai
Pillow==10.0.0
xhtml2pdf
emoji
streamlit-chat-widget